+++ packages/@aws-cdk/aws-applicationautoscaling/test/test.cron.ts
@@ -1,4 +1,4 @@
-import { Duration } from '@aws-cdk/core';
+import { Duration, Stack, Lazy } from '@aws-cdk/core';
 import { Test } from 'nodeunit';
 import * as appscaling from '../lib';
 
@@ -15,8 +15,15 @@ export = {
 
   'rate must be whole number of minutes'(test: Test) {
     test.throws(() => {
-      appscaling.Schedule.rate(Duration.seconds(12345));
-    }, /'12345 seconds' cannot be converted into a whole number of minutes/);
+      appscaling.Schedule.rate(Duration.minutes(0.13456));
+    }, /'0.13456 minutes' cannot be converted into a whole number of seconds/);
+    test.done();
+  },
+
+  'rate must not be in seconds'(test: Test) {
+    test.throws(() => {
+      appscaling.Schedule.rate(Duration.seconds(1));
+    }, /Allowed unit for scheduling is: 'minute', 'minutes', 'hour', 'hours', 'day' or 'days'/);
     test.done();
   },
 
@@ -24,6 +31,20 @@ export = {
     test.throws(() => {
       appscaling.Schedule.rate(Duration.days(0));
     }, /Duration cannot be 0/);
+    test.done();
+  },
+
+  'rate can be token'(test: Test) {
+    const stack = new Stack();
+    const lazyDuration = Duration.minutes(Lazy.number({ produce: () => 5 }));
+    const rate = appscaling.Schedule.rate(lazyDuration);
+    test.equal('rate(5 minutes)', stack.resolve(rate).expressionString);
+    test.done();
+  },
+
+  'rate can be in allowed type hours'(test: Test) {
+    test.equal('rate(1 hour)', appscaling.Schedule.rate(Duration.hours(1))
+      .expressionString);
     test.done();
   },
 };