+++ packages/@aws-cdk/aws-elasticloadbalancingv2/lib/alb/application-listener.ts
@@ -266,9 +266,7 @@ export class ApplicationListener extends BaseListener implements IApplicationLis
     // Only one certificate can be specified per resource, even though
     // `certificates` is of type Array
     for (let i = 0; i < additionalCerts.length; i++) {
-      // ids should look like: `id`, `id2`, `id3` (for backwards-compatibility)
-      const certId = (i > 0) ? `${id}${i + 1}` : id;
-      new ApplicationListenerCertificate(this, certId, {
+      new ApplicationListenerCertificate(this, `${id}${i + 1}`, {
         listener: this,
         certificates: [additionalCerts[i]],
       });